requirements.txt
beautifulsoup4==4.12.2
requests==2.28.1
streamlink==6.3.1
websocket-client==1.6.3
nose3==1.3.8
